Request duplicate header fields are catenated
It is legal to have multiple header field keys only if those values are
safe to catenate with comma separators without changing the semantics of
the message. Since we only deal with single header fields within Served
we now do the catenating of duplicates as part of the parse process.

There is scope in the future for having helpers to split those headers
by comma, and also to merge multiple header fields of a response rather
than replace using a new syntax.

Ref: http://www.w3.org/Protocols/rfc2616/rfc2616-sec4.html#sec4.2